+++ src/game/Game.java
@@ -25,8 +25,8 @@ import game.UI.elements.containers.BasicUIContainer;
 import game.UI.elements.image.UIImage;
 import game.UI.elements.text.UILabel;
 import game.debug.IDHandlerDebugFrame;
-import game.debug.logTool.Log;
-import game.debug.logTool.frame.LogFrame;
+import game.debug.log.Log;
+import game.debug.log.frame.LogFrame;
 import game.gameObject.GameObject;
 import game.gameObject.graphics.Camera;
 import game.gameObject.graphics.Paintable;
